Module file for encryption/decryption methods
